+++ airbyte-cdk/python/unit_tests/sources/declarative/retrievers/test_simple_retriever.py
@@ -2,28 +2,31 @@
 # Copyright (c) 2022 Airbyte, Inc., all rights reserved.
 #
 
-from unittest.mock import MagicMock
+from unittest.mock import MagicMock, patch
 
 import airbyte_cdk.sources.declarative.requesters.error_handlers.response_status as response_status
 import pytest
 import requests
 from airbyte_cdk.models import SyncMode
-from airbyte_cdk.sources.declarative.read_exception import ReadException
+from airbyte_cdk.sources.declarative.exceptions import ReadException
 from airbyte_cdk.sources.declarative.requesters.error_handlers.response_action import ResponseAction
 from airbyte_cdk.sources.declarative.requesters.error_handlers.response_status import ResponseStatus
 from airbyte_cdk.sources.declarative.requesters.request_option import RequestOptionType
 from airbyte_cdk.sources.declarative.requesters.requester import HttpMethod
 from airbyte_cdk.sources.declarative.retrievers.simple_retriever import SimpleRetriever
+from airbyte_cdk.sources.streams.http.auth import NoAuth
+from airbyte_cdk.sources.streams.http.http import HttpStream
 
 primary_key = "pk"
 records = [{"id": 1}, {"id": 2}]
 config = {}
 
 
-def test_simple_retriever_full():
+@patch.object(HttpStream, "read_records", return_value=[])
+def test_simple_retriever_full(mock_http_stream):
     requester = MagicMock()
     request_params = {"param": "value"}
-    requester.request_params.return_value = request_params
+    requester.get_request_params.return_value = request_params
 
     paginator = MagicMock()
     next_page_token = {"cursor": "cursor_value"}
@@ -42,6 +45,7 @@ def test_simple_retriever_full():
     underlying_state = {"date": "2021-01-01"}
     iterator.get_stream_state.return_value = underlying_state
 
+    requester.get_authenticator.return_value = NoAuth
     url_base = "https://airbyte.io"
     requester.get_url_base.return_value = url_base
     path = "/v1"
@@ -51,10 +55,13 @@ def test_simple_retriever_full():
     backoff_time = 60
     should_retry = ResponseStatus.retry(backoff_time)
     requester.should_retry.return_value = should_retry
-    request_body_data = {"body": "data"}
-    requester.request_body_data.return_value = request_body_data
     request_body_json = {"body": "json"}
     requester.request_body_json.return_value = request_body_json
+
+    request_body_data = {"body": "data"}
+    requester.get_request_body_data.return_value = request_body_data
+    request_body_json = {"body": "json"}
+    requester.get_request_body_json.return_value = request_body_json
     request_kwargs = {"kwarg": "value"}
     requester.request_kwargs.return_value = request_kwargs
     cache_filename = "cache"
@@ -63,12 +70,13 @@ def test_simple_retriever_full():
     requester.use_cache = use_cache
 
     retriever = SimpleRetriever(
-        "stream_name",
-        primary_key,
+        name="stream_name",
+        primary_key=primary_key,
         requester=requester,
         paginator=paginator,
         record_selector=record_selector,
         stream_slicer=iterator,
+        options={},
     )
 
     assert retriever.primary_key == primary_key
@@ -89,11 +97,13 @@ def test_simple_retriever_full():
     assert not retriever.raise_on_http_errors
     assert retriever.should_retry(requests.Response())
     assert retriever.backoff_time(requests.Response()) == backoff_time
-    assert retriever.request_body_data(None, None, None) == request_body_data
     assert retriever.request_body_json(None, None, None) == request_body_json
     assert retriever.request_kwargs(None, None, None) == request_kwargs
     assert retriever.cache_filename == cache_filename
     assert retriever.use_cache == use_cache
+
+    [r for r in retriever.read_records(SyncMode.full_refresh)]
+    paginator.reset.assert_called()
 
 
 @pytest.mark.parametrize(
@@ -106,7 +116,7 @@ def test_simple_retriever_full():
 )
 def test_should_retry(test_name, requester_response, expected_should_retry, expected_backoff_time):
     requester = MagicMock()
-    retriever = SimpleRetriever("stream_name", primary_key, requester=requester, record_selector=MagicMock())
+    retriever = SimpleRetriever(name="stream_name", primary_key=primary_key, requester=requester, record_selector=MagicMock(), options={})
     requester.should_retry.return_value = requester_response
     assert retriever.should_retry(requests.Response()) == expected_should_retry
     if requester_response.action == ResponseAction.RETRY:
@@ -125,7 +135,9 @@ def test_parse_response(test_name, status_code, response_status, len_expected_re
     requester = MagicMock()
     record_selector = MagicMock()
     record_selector.select_records.return_value = [{"id": 100}]
-    retriever = SimpleRetriever("stream_name", primary_key, requester=requester, record_selector=record_selector)
+    retriever = SimpleRetriever(
+        name="stream_name", primary_key=primary_key, requester=requester, record_selector=record_selector, options={}
+    )
     response = requests.Response()
     response.status_code = status_code
     requester.should_retry.return_value = response_status
@@ -154,7 +166,9 @@ def test_backoff_time(test_name, response_action, retry_in, expected_backoff_tim
     record_selector = MagicMock()
     record_selector.select_records.return_value = [{"id": 100}]
     response = requests.Response()
-    retriever = SimpleRetriever("stream_name", primary_key, requester=requester, record_selector=record_selector)
+    retriever = SimpleRetriever(
+        name="stream_name", primary_key=primary_key, requester=requester, record_selector=record_selector, options={}
+    )
     if expected_backoff_time:
         requester.should_retry.return_value = ResponseStatus(response_action, retry_in)
         actual_backoff_time = retriever.backoff_time(response)
@@ -168,35 +182,87 @@ def test_backoff_time(test_name, response_action, retry_in, expected_backoff_tim
 
 
 @pytest.mark.parametrize(
-    "test_name, paginator_mapping, expected_mapping",
+    "test_name, paginator_mapping, stream_slicer_mapping, expected_mapping",
     [
-        ("test_only_base_headers", {}, {"key": "value"}),
-        ("test_header_from_pagination", {"offset": 1000}, {"key": "value", "offset": 1000}),
-        ("test_duplicate_header", {"key": 1000}, None),
+        ("test_only_base_headers", {}, {}, {"key": "value"}),
+        ("test_header_from_pagination", {"offset": 1000}, {}, {"key": "value", "offset": 1000}),
+        ("test_header_from_stream_slicer", {}, {"slice": "slice_value"}, {"key": "value", "slice": "slice_value"}),
+        ("test_duplicate_header_slicer", {}, {"key": "slice_value"}, None),
+        ("test_duplicate_header_slicer_paginator", {"k": "v"}, {"k": "slice_value"}, None),
+        ("test_duplicate_header_paginator", {"key": 1000}, {}, None),
     ],
 )
-def test_get_request_options_from_pagination(test_name, paginator_mapping, expected_mapping):
+def test_get_request_options_from_pagination(test_name, paginator_mapping, stream_slicer_mapping, expected_mapping):
+    # This test does not test request headers because they must be strings
     paginator = MagicMock()
-    paginator.request_headers.return_value = paginator_mapping
-    paginator.request_params.return_value = paginator_mapping
-    paginator.request_body_data.return_value = paginator_mapping
-    paginator.request_body_json.return_value = paginator_mapping
-    requester = MagicMock()
+    paginator.get_request_params.return_value = paginator_mapping
+    paginator.get_request_body_data.return_value = paginator_mapping
+    paginator.get_request_body_json.return_value = paginator_mapping
+
+    stream_slicer = MagicMock()
+    stream_slicer.get_request_params.return_value = stream_slicer_mapping
+    stream_slicer.get_request_body_data.return_value = stream_slicer_mapping
+    stream_slicer.get_request_body_json.return_value = stream_slicer_mapping
 
     base_mapping = {"key": "value"}
-    requester.request_headers.return_value = base_mapping
-    requester.request_params.return_value = base_mapping
-    requester.request_body_data.return_value = base_mapping
-    requester.request_body_json.return_value = base_mapping
+    requester = MagicMock()
+    requester.get_request_params.return_value = base_mapping
+    requester.get_request_body_data.return_value = base_mapping
+    requester.get_request_body_json.return_value = base_mapping
 
     record_selector = MagicMock()
-    retriever = SimpleRetriever("stream_name", primary_key, requester=requester, record_selector=record_selector, paginator=paginator)
+    retriever = SimpleRetriever(
+        name="stream_name",
+        primary_key=primary_key,
+        requester=requester,
+        record_selector=record_selector,
+        paginator=paginator,
+        stream_slicer=stream_slicer,
+        options={},
+    )
 
     request_option_type_to_method = {
-        RequestOptionType.header: retriever.request_headers,
         RequestOptionType.request_parameter: retriever.request_params,
         RequestOptionType.body_data: retriever.request_body_data,
         RequestOptionType.body_json: retriever.request_body_json,
+    }
+
+    for _, method in request_option_type_to_method.items():
+        if expected_mapping:
+            actual_mapping = method(None, None, None)
+            assert expected_mapping == actual_mapping
+        else:
+            try:
+                method(None, None, None)
+                assert False
+            except ValueError:
+                pass
+
+
+@pytest.mark.parametrize(
+    "test_name, paginator_mapping, expected_mapping",
+    [
+        ("test_only_base_headers", {}, {"key": "value"}),
+        ("test_header_from_pagination", {"offset": 1000}, {"key": "value", "offset": "1000"}),
+        ("test_duplicate_header", {"key": 1000}, None),
+    ],
+)
+def test_get_request_headers(test_name, paginator_mapping, expected_mapping):
+    # This test is separate from the other request options because request headers must be strings
+    paginator = MagicMock()
+    paginator.get_request_headers.return_value = paginator_mapping
+    requester = MagicMock()
+
+    base_mapping = {"key": "value"}
+    requester.get_request_headers.return_value = base_mapping
+
+    record_selector = MagicMock()
+    retriever = SimpleRetriever(
+        name="stream_name", primary_key=primary_key, requester=requester, record_selector=record_selector, paginator=paginator, options={}
+    )
+
+    request_option_type_to_method = {
+        RequestOptionType.header: retriever.request_headers,
     }
 
     for _, method in request_option_type_to_method.items():
@@ -223,13 +289,20 @@ def test_get_request_options_from_pagination(test_name, paginator_mapping, expec
 )
 def test_request_body_data(test_name, requester_body_data, paginator_body_data, expected_body_data):
     paginator = MagicMock()
-    paginator.request_body_data.return_value = paginator_body_data
+    paginator.get_request_body_data.return_value = paginator_body_data
     requester = MagicMock()
 
-    requester.request_body_data.return_value = requester_body_data
+    requester.get_request_body_data.return_value = requester_body_data
 
     record_selector = MagicMock()
-    retriever = SimpleRetriever("stream_name", primary_key, requester=requester, record_selector=record_selector, paginator=paginator)
+    retriever = SimpleRetriever(
+        name="stream_name",
+        primary_key=primary_key,
+        requester=requester,
+        record_selector=record_selector,
+        paginator=paginator,
+        options={},
+    )
 
     if expected_body_data:
         actual_body_data = retriever.request_body_data(None, None, None)
@@ -257,7 +330,14 @@ def test_path(test_name, requester_path, paginator_path, expected_path):
     requester.get_path.return_value = requester_path
 
     record_selector = MagicMock()
-    retriever = SimpleRetriever("stream_name", primary_key, requester=requester, record_selector=record_selector, paginator=paginator)
+    retriever = SimpleRetriever(
+        name="stream_name",
+        primary_key=primary_key,
+        requester=requester,
+        record_selector=record_selector,
+        paginator=paginator,
+        options={},
+    )
 
     actual_path = retriever.path(stream_state=None, stream_slice=None, next_page_token=None)
     assert expected_path == actual_path